python: split router from bridge
